pyproject.toml
[tool.poetry]
name = "test-bot"
version = "0.1.0"
description = ""
authors = ["khasanjon <khasanjon.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.2.0"
python-dotenv = "^1.0.0"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
